pyproject.toml
[project]
name = "agent2agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.5.5",
    "fastapi>=0.115.12",
    "google-adk>=0.1.0",
    "langchain-huggingface>=0.1.2",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.51",
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.12",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
    "weaviate>=0.1.2",
    "loguru>=0.7.3",
    "langchain-redis>=0.2.0",
    "redis>=5.2.1",
    "langfuse>=2.60.3",
]
